:recycle: refactor(theme)!: migrate cast/tome palettes to the shared sill catalog (atelier block-6) (#141)
* :recycle: refactor(theme)!: migrate to sill 0.3.0 — derive cast/tome palettes from ThemeSpec (atelier block-6)

wand's cast HUD + tome launcher palettes are now DERIVED from sill's
authoritative ThemeSpec catalog instead of hand-kept per-theme literals
— "stop saying copy facet's theme" now holds in code. One sill catalog
change reaches both wand surfaces.

WandCore → Palette, WandAdapterMacOS → Palette + Effects (NOT PaletteKit:
wand has its own NSColorParse and never uses pal/resolve, like perch).

Theme model: enum CastTheme/TomeTheme → String names validated against
sill canonicalThemeNames (+ wand-local engine themes). wandCastPalette /
wandTomePalette project sill roles onto wand's String-token surfaces
(maximal sill-driven, atelier Q5): primary→trail, error→noMatch,
foreground→text, background→cards/outline/badge.

Catalog (atelier Q1/Q6, breaking): wand adopts sill's full 12-theme
catalog (gains cobalt2/shades-of-purple/tokyo-hack/github-dark/
github-light/catppuccin-latte) + system. Cut: nord/onedark/rosepine/
vapor/aurora/mono/bare-catppuccin/default → clamp to the native `system`
default (loud log). terminal local green override dropped — sill terminal
is now green-on-black (Phase V Q4). Default theme: default → system.

Engine themes kept wand-local (no sill palette): neon (per-frame neon
cycle), splatoon (per-stroke Turf-War ink). chomp palette derives its
arcade constants from sill chomp ThemeSpec (byte-match, zero visual
change) but keeps its wall-blue arrangement + ChompRenderer geometry.

Dedup (atelier Q4): NSColorParse.neonInks deleted → derived from sill
EffectSpec.neon.flash (byte-match). wand keeps its own per-frame cycle
engine — only the DATA is shared.

Intensity (atelier Q3/Q9): wand's local Intensity enum removed; consumes
sill EffectIntensity (rule-of-three with perch met). Adapter wraps the
Double multiplier as CGFloat at the use site.

No functional change to gesture/launcher behavior — theme palettes only.
config.toml theme docs rewritten to the new catalog. ThemeTests added.

DURING migration sill is a path dep (atomic co-edit for EffectIntensity);
the release commit swaps to url + sill 0.4.0.


* :bug: fix(theme): correct stale system-default test + docs after sill migration (review fixes)

Adversarial review (wf_2efb8851) blockers:
- ConfigTests.testOverlayDefaultsWhenAbsent asserted the dead #3b82f6
  default; unset theme now resolves to system → trail token "accent".
- config.toml + Config.swift + GestureOverlay.swift comments still said
  theme "default" (a cut name); the canonical default is "system".
- Main.swift burst intensity wrapped CGFloat() to match GestureOverlay
  (EffectIntensity.multiplier is Double).

Tome hover-text contrast (5 mid-luminance themes) is fixed upstream in
sill bestForeground (WCAG); no wand bridge change needed.


* :package: build(deps): pin sill 0.4.0 via url (replace migration path-dep)

Swap the block-6 path dep (.package(path:"../sill")) for the released
url dep .upToNextMinor(from:"0.4.0") and track Package.resolved (locks
sill at 2577a09 / 0.4.0) so CI + brew builds are reproducible. Removes
Package.resolved from .gitignore (wand is now a sill consumer).


---------

+++ Sources/WandApp/Main.swift
@@ -502,7 +502,7 @@ enum WandApp {
                     burstManager.emit(
                         at: cocoaPoint, color: burstColor,
                         kind: burstSpec.kind,
-                        intensity: cfg.intensity.multiplier)
+                        intensity: CGFloat(cfg.intensity.multiplier))
                     arcadeScoreManager.emit(
                         at: cocoaPoint, color: burstColor,
                         kind: burstSpec.kind)
